refactor: 코드 전반적인 리팩토링
- 코드 일관성을 유지하기 위해 일부 수정
- ""의 하드코딩을 constant로 관리
- Notification의 Databinding
- customview들을 한 패키지로 모아서 관리
