+++ app/src/main/java/com/github/repo/presentation/common/Constant.kt
@@ -1,0 +1,5 @@
+package com.github.repo.presentation.common
+
+object Constant {
+    const val BLANK = ""
+}
