+++ app/src/main/java/com/github/repo/presentation/common/custom/CustomToolbar.kt
@@ -1,4 +1,4 @@
-package com.github.repo.presentation.common
+package com.github.repo.presentation.common.custom
 
 import android.content.Context
 import android.util.AttributeSet
