+++ app/src/main/res/layout/fragment_profile.xml
@@ -17,7 +17,7 @@
         android:background="@drawable/background_main"
         tools:context=".presentation.profile.ProfileFragment">
 
-        <com.github.repo.presentation.common.CustomToolbar
+        <com.github.repo.presentation.common.custom.CustomToolbar
             android:id="@+id/tb_profile"
             android:layout_width="0dp"
             android:layout_height="72dp"
@@ -126,7 +126,7 @@
                 android:orientation="vertical"
                 app:layout_constraintTop_toBottomOf="@id/divider">
 
-                <com.github.repo.presentation.profile.custom.CustomProfileUserInfoItem
+                <com.github.repo.presentation.common.custom.CustomProfileUserInfoItem
                     android:id="@+id/item_live_place"
                     android:layout_width="match_parent"
                     android:layout_height="34dp"
@@ -135,7 +135,7 @@
                     app:contentText='@{profile.location ==  null ? "": profile.location}'
                     app:itemType="text" />
 
-                <com.github.repo.presentation.profile.custom.CustomProfileUserInfoItem
+                <com.github.repo.presentation.common.custom.CustomProfileUserInfoItem
                     android:id="@+id/item_blog"
                     android:layout_width="match_parent"
                     android:layout_height="34dp"
@@ -144,7 +144,7 @@
                     app:contentText='@{profile.blogUrl ==  null ? "": profile.blogUrl}'
                     app:itemType="link" />
 
-                <com.github.repo.presentation.profile.custom.CustomProfileUserInfoItem
+                <com.github.repo.presentation.common.custom.CustomProfileUserInfoItem
                     android:id="@+id/item_mail"
                     android:layout_width="match_parent"
                     android:layout_height="34dp"
@@ -153,7 +153,7 @@
                     app:contentText='@{profile.email ==  null ? "": profile.email}'
                     app:itemType="mail" />
 
-                <com.github.repo.presentation.profile.custom.CustomProfileUserInfoItem
+                <com.github.repo.presentation.common.custom.CustomProfileUserInfoItem
                     android:id="@+id/item_friends"
                     android:layout_width="match_parent"
                     android:layout_height="34dp"
@@ -173,7 +173,7 @@
                 android:paddingBottom="8dp"
                 app:layout_constraintTop_toBottomOf="@+id/linearLayout3">
 
-                <com.github.repo.presentation.profile.custom.CustomProfileGithubDataItem
+                <com.github.repo.presentation.common.custom.CustomProfileGithubDataItem
                     android:id="@+id/item_repository"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -181,7 +181,7 @@
                     app:icon="@drawable/ic_bookmark"
                     app:name="Repositories" />
 
-                <com.github.repo.presentation.profile.custom.CustomProfileGithubDataItem
+                <com.github.repo.presentation.common.custom.CustomProfileGithubDataItem
                     android:id="@+id/item_organization"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -189,7 +189,7 @@
                     app:icon="@drawable/ic_users"
                     app:name="Organizations" />
 
-                <com.github.repo.presentation.profile.custom.CustomProfileGithubDataItem
+                <com.github.repo.presentation.common.custom.CustomProfileGithubDataItem
                     android:id="@+id/item_starred"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
