+++ app/src/main/res/layout/item_notification.xml
@@ -1,14 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
+        <variable
+            name="notification"
+            type="com.github.repo.domain.model.Notification" />
     </data>
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
+
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/erase_item_view"
             android:layout_width="match_parent"
@@ -18,18 +21,18 @@
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="110dp"
                 android:layout_height="match_parent"
+                app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent">
+                app:layout_constraintTop_toTopOf="parent">
 
                 <ImageView
                     android:layout_width="24dp"
                     android:layout_height="24dp"
-                    app:layout_constraintTop_toTopOf="parent"
+                    android:src="@drawable/ic_check"
                     app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    android:src="@drawable/ic_check" />
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent" />
             </androidx.constraintlayout.widget.ConstraintLayout>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
@@ -38,8 +41,8 @@
             android:id="@+id/layout_swipe"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:background="@color/navy_light">
+            android:background="@color/navy_light"
+            android:orientation="vertical">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
@@ -61,6 +64,7 @@
 
                     <ImageView
                         android:id="@+id/iv_thumbnail"
+                        setImage="@{notification.thumbnailUrl}"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
                         android:src="@drawable/ic_user" />
@@ -68,53 +72,53 @@
 
                 <TextView
                     android:id="@+id/tv_repo_name"
+                    style="@style/SubTitle2.grey"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:maxWidth="200dp"
                     android:layout_marginStart="8dp"
-                    android:text="sam/whatsinmygithub"
-                    android:textSize="16sp"
-                    android:maxLines="1"
                     android:ellipsize="end"
-                    style="@style/SubTitle2.grey"
+                    android:maxWidth="200dp"
+                    android:maxLines="1"
+                    android:text="@{notification.repoName}"
+                    android:textSize="16sp"
                     app:layout_constraintBottom_toBottomOf="@+id/cardView"
                     app:layout_constraintStart_toEndOf="@+id/cardView"
                     app:layout_constraintTop_toTopOf="@+id/cardView" />
 
                 <TextView
                     android:id="@+id/tv_notification_title"
+                    style="@style/SubTitle2.white"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="8dp"
                     android:layout_marginTop="8dp"
-                    android:maxLines="1"
                     android:ellipsize="end"
                     android:maxWidth="240dp"
-                    android:text="프로젝트 요구사항 및 참고자료"
+                    android:maxLines="1"
+                    android:text="@{notification.notificationTitle}"
                     android:textSize="16sp"
-                    style="@style/SubTitle2.white"
                     app:layout_constraintStart_toEndOf="@+id/cardView"
                     app:layout_constraintTop_toBottomOf="@id/tv_repo_name" />
 
                 <TextView
                     android:id="@+id/tv_issue_number"
+                    style="@style/SubTitle2.grey"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="8dp"
-                    android:text="#1"
+                    android:text='@{"#"+notification.threadId}'
                     android:textSize="16sp"
-                    style="@style/SubTitle2.grey"
-                    app:layout_constraintStart_toEndOf="@+id/tv_repo_name"
                     app:layout_constraintBottom_toBottomOf="@+id/cardView"
+                    app:layout_constraintStart_toEndOf="@+id/tv_repo_name"
                     app:layout_constraintTop_toTopOf="@+id/cardView" />
 
                 <TextView
                     android:id="@+id/tv_update_time"
+                    style="@style/SubTitle2.grey"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="1일 전"
+                    android:text="@{notification.updateTime}"
                     android:textSize="16sp"
-                    style="@style/SubTitle2.grey"
                     app:layout_constraintBottom_toBottomOf="@+id/cardView"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintTop_toTopOf="@+id/cardView" />
@@ -135,9 +139,9 @@
                         android:layout_height="match_parent"
                         android:gravity="center"
                         android:src="@drawable/ic_user"
+                        android:text="@{notification.commentCount}"
                         android:textColor="@color/primary"
-                        android:textSize="12sp"
-                        android:text="1" />
+                        android:textSize="12sp" />
                 </androidx.cardview.widget.CardView>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
@@ -145,8 +149,8 @@
             <View
                 android:layout_width="match_parent"
                 android:layout_height="0.5dp"
-                android:background="@color/grey"
-                android:layout_marginStart="24dp" />
+                android:layout_marginStart="24dp"
+                android:background="@color/grey" />
         </LinearLayout>
     </FrameLayout>
 </layout>
